+++ src/VaultSharp/V1/AuthMethods/AppRole/Models/SecretIdInfo.cs
@@ -9,10 +9,13 @@ namespace VaultSharp.V1.AuthMethods.AppRole.Models
     {
         [JsonProperty("secret_id_accessor")]
         public string SecretIdAccessor { get; set; }
+        
         [JsonProperty("secret_id")]
         public string SecretId { get; set; }
+        
         [JsonProperty("secret_id_ttl")]
         public int SecretIdTimeToLive { get; set; }
+        
         [JsonProperty("secret_id_num_uses")]
         public int SecretIdNumUses { get; set; }
     }