+++ src/VaultSharp/V1/AuthMethods/Okta/OktaTOTPProvider.cs
@@ -1,0 +1,11 @@
+﻿namespace VaultSharp.V1.AuthMethods.Okta
+{
+    /// <summary>
+    /// The OktaTOTPProvider
+    /// </summary>
+    public enum OktaTOTPProvider
+    {
+        OKTA,
+        GOOGLE
+    }
+}