+++ src/VaultSharp/V1/AuthMethods/Okta/OktaVerifyPushChallengeResponse.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.AuthMethods.Okta
+{
+    public class OktaVerifyPushChallengeResponse
+    {
+        /// <summary>
+        /// The correct push challenge answer
+        /// </summary>
+        [JsonProperty("correct_answer")]
+        public int CorrectAnswer { get; set; }
+    }
+}